solubility/requirements.txt
fastapi
uvicorn
python-json-logger
mypy
pydantic
numpy
requests
types-requests
fair-esm
--extra-index-url https://download.pytorch.org/whl/cu118 torch==2.0.1

